# templates/base-template/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AgentInfoModel = {
  preferred: string;
  fallback?: string[];
};

export type AgentInfoCompliance = {
  riskLevel: string;
  humanInTheLoop: boolean;
  dataClassification: string;
  regulatoryFrameworks: string[];
};

export interface AgentInfo {
  name: string;
  version: string;
  description: string;
  model: AgentInfoModel;
  tools: string[];
  skillNames: string[];
  compliance: AgentInfoCompliance;
  soulContent: string;
  rulesContent: string;
  dutiesContent: string;
}

export interface AgentSkill {
  name: string;
  description: string;
  status: string;
  lastUsed?: string;
}

export type AgentSkillDetailScriptsItem = {
  name: string;
  content: string;
};

export interface AgentSkillDetail {
  name: string;
  description: string;
  content: string;
  scripts: AgentSkillDetailScriptsItem[];
}

export interface FileTreeNode {
  name: string;
  path: string;
  type: string;
  children?: FileTreeNode[];
}

export interface FileContent {
  path: string;
  content: string;
  language: string;
}

export type ComplianceStatusRegulatoryFrameworksItem = {
  name: string;
  status: string;
  lastValidated?: string;
};

export type ComplianceStatusValidationChecksItem = {
  name: string;
  status: string;
  nextDue?: string;
};

export interface ComplianceStatus {
  riskLevel: string;
  humanInTheLoop: boolean;
  dataClassification: string;
  regulatoryFrameworks: ComplianceStatusRegulatoryFrameworksItem[];
  auditLogging: boolean;
  retentionDays: number;
  reviewCadence: string;
  approvers: string[];
  validationChecks?: ComplianceStatusValidationChecksItem[];
}

export type AgentChatRequestMessagesItem = {
  role: string;
  content: string;
};

export interface AgentChatRequest {
  messages: AgentChatRequestMessagesItem[];
  sessionId?: string;
}

export interface DashboardMetric {
  id: string;
  label: string;
  value: string;
  change: string;
  changeType: string;
  trend: string;
  detail?: string;
}

export interface DashboardInsight {
  id: string;
  severity: string;
  headline: string;
  summary: string;
  category: string;
  actionLabel?: string;
}

export interface ActivityEntry {
  id: string;
  timestamp: string;
  skill: string;
  action: string;
  status: string;
}

export interface ApiError {
  error: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type GetAgentFileContentParams = {
  /**
   * Relative file path within the agent directory
   */
  path: string;
};