# templates/base-template/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agent/skills/{name}:
    get:
      operationId: getAgentSkill
      tags: [agent]
      summary: Get full skill definition
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skill detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSkillDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /agent/file:
    get:
      operationId: getAgentFileContent
      tags: [agent]
      summary: Get file content by path query parameter
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Relative file path within the agent directory
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "400":
          description: Missing path parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /agent/chat:
    post:
      operationId: agentChat
      tags: [agent]
      summary: Chat with agent (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentChatRequest"
      responses:
        "200":
          description: SSE stream of agent response
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    AgentSkillDetail:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        content:
          type: string
        scripts:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              content:
                type: string
            required:
              - name
              - content
      required:
        - name
        - description
        - content
        - scripts
...
    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        language:
          type: string
      required:
        - path
        - content
        - language
...
    AgentChatRequest:
      type: object
      properties:
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
            required:
              - role
              - content
        sessionId:
          type: string
      required:
        - messages
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...