# templates/base-template/lib/api-zod/src/generated/types/agentChatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentChatRequestMessagesItem } from "./agentChatRequestMessagesItem";

export interface AgentChatRequest {
  messages: AgentChatRequestMessagesItem[];
  sessionId?: string;
}