# templates/base-template/lib/api-zod/src/generated/types/agentSkillDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentSkillDetailScriptsItem } from "./agentSkillDetailScriptsItem";

export interface AgentSkillDetail {
  name: string;
  description: string;
  content: string;
  scripts: AgentSkillDetailScriptsItem[];
}